Split glucoseGames.ts into smaller files
