+++ utils/formatting/percentFormatting.ts
@@ -1,0 +1,4 @@
+export const cleanPercentForDisplay = (percentTimeInRange: number) => {
+  if (!percentTimeInRange) return '0.00'
+  return percentTimeInRange.toFixed(2)
+}
