+++ utils/games/averageInRange/averageInRangeGames.ts
@@ -4,7 +4,8 @@ import { scoreRecordsByAverageGlucose } from '~/utils/scoring/averageGlucoseValu
 import { filterRecordsByTimePeriod } from '~/utils/filters/timePeriod/filterByTimePeriod'
 import type { ScoredDay } from '~/types/scoredDay'
 import { CurrentDayStatus } from '~/types/constants'
-import { calculateDailyStreakStats } from '~/utils/glucoseGames'
+import { calculateDailyStreakStats } from '~/utils/streaks/dailyStreaks'
+import { cleanPercentForDisplay } from '~/utils/formatting/percentFormatting'
 
 export const averageInRangeGame = (
   records: GlucoseRecord[],
